# api-specs.yaml
openapi: 3.0.3
info:
  title: Shadow Ledger System API
  description: Banking backend system for processing financial events and maintaining shadow ledger
  version: 1.0.0
  contact:
    name: Shadow Ledger Team
servers:
  - url: http://localhost:8080
    description: API Gateway
  - url: http://localhost:8081
    description: Event Service (Direct)
  - url: http://localhost:8082
    description: Shadow Ledger Service (Direct)
  - url: http://localhost:8083
    description: Drift Correction Service (Direct)
...
paths:
...
  /drift-check:
    post:
      tags:
        - Drift Correction
      summary: Check for balance drift
      description: Compare CBS balances with shadow ledger
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CbsBalance'
            example:
              - accountId: A10
                reportedBalance: 800
              - accountId: A11
                reportedBalance: 1500
      responses:
        '200':
          description: Drift check completed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CorrectionEvent'
              example:
                - eventId: CORR-A10-12345
                  accountId: A10
                  type: credit
                  amount: 50
  /correct/{accountId}:
    post:
      tags:
        - Drift Correction
      summary: Manual correction
      description: Create manual correction event
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
          example: A10
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [credit, debit]
          example: credit
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 100
      responses:
        '200':
          description: Correction event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionEvent'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CbsBalance:
      type: object
      required:
        - accountId
        - reportedBalance
      properties:
        accountId:
          type: string
          example: A10
        reportedBalance:
          type: number
          format: double
          example: 800
    CorrectionEvent:
      type: object
      properties:
        eventId:
          type: string
          example: CORR-A10-12345
        accountId:
          type: string
          example: A10
        type:
          type: string
          enum: [credit, debit]
          example: credit
        amount:
          type: number
          format: double
          example: 50
...